refactor: fix the native styles
